# Switch to godirwalk for a bit better performance