--- a/go.mod
+++ b/go.mod
@@ -3,6 +3,7 @@ module github.com/gadget-inc/fsdiff
 go 1.16
 
 require (
+	github.com/karrick/godirwalk v1.16.1
 	github.com/klauspost/compress v1.13.6
 	github.com/minio/sha256-simd v1.0.0
 	google.golang.org/protobuf v1.26.0