- It doesn't throw errors when trying to stat files that have just been removed
 - It saves us one syscall per ignored file
 - It is more memory efficient

Don't see why not!

On my machine on `main`:

```
cd test && go test -bench=. -benchtime=10x
goos: darwin
goarch: amd64
pkg: github.com/gadget-inc/fsdiff/test
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkSimpleInitialDiff-12    	      10	    413688 ns/op
BenchmarkReactInitialDiff-12     	      10	2493167343 ns/op
BenchmarkReactChangedDiff-12     	      10	3578962576 ns/op
```

on this branch:

```
goos: darwin
goarch: amd64
pkg: github.com/gadget-inc/fsdiff/test
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkSimpleInitialDiff-12    	      10	    742025 ns/op
BenchmarkReactInitialDiff-12     	      10	2632499614 ns/op
BenchmarkReactChangedDiff-12     	      10	3750612520 ns/op
PASS
```

Doesn't really seem much faster but isn't slower and saves us one more error condition so I think it's still worth it.