@airhorns 2021-11-27
I read those results wrong, bigger numbers are worse and godirwalk has bigger numbers across the board! Alex mentioned that go 1.16 added the stdlib function we're using to make one less call in the same way it does, so let's just stick with that for now and change one thing at a time.